pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "medium-notion-translator"
version = "0.1.0"
description = "Medium記事を日本語に翻訳してNotionに追加するCLIツール"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Akira Fujita"}
]

dependencies = [
    "click>=8.0",
    "python-dotenv>=1.0",
    "playwright>=1.40",
    "notion-client>=2.0",
    "pydantic>=2.0",
    "loguru>=0.7",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
]

[project.scripts]
medium-notion = "medium_notion.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
